check missing of translation key also in alert-error.component and move check to utils
