+++ src/main/webapp/app/shared/alert/alert.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnDestroy, OnInit } from '@angular/core';
 import { JhiAlert } from 'ng-jhipster';
 import { JhiAlertService } from 'ng-jhipster';
-import * as Sentry from '@sentry/browser';
+import { checkForMissingTranslationKey } from 'app/shared/util/utils';
 
 @Component({
     selector: 'jhi-alert',
@@ -28,18 +28,7 @@ export class AlertComponent implements OnInit, OnDestroy {
         // This is a workaround to avoid translation not found issues.
         if (this.alerts && Array.isArray(this.alerts)) {
             this.alerts.forEach((alert) => {
-                if (alert?.msg?.startsWith('translation-not-found')) {
-                    // In case a translation key is not found, remove the 'translation-not-found[...]' annotation
-                    const alertMessageMatch = alert.msg.match(/translation-not-found\[(.*?)\]$/);
-                    if (alertMessageMatch && alertMessageMatch.length > 1) {
-                        alert.msg = alertMessageMatch[1];
-                    } else {
-                        // Fallback, in case the bracket is missing
-                        alert.msg = alert.msg.replace('translation-not-found', '');
-                    }
-                    // Sent a sentry warning with the translation key
-                    Sentry.captureException(new Error('Unknown translation key: ' + alert.msg));
-                }
+                checkForMissingTranslationKey(alert);
             });
         }
     }
